% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{core_packages}
\alias{core_packages}
\title{Core R4SUB packages}
\usage{
core_packages()
}
\value{
A character vector of package names.
}
\description{
Returns the names of the R4SUB packages that are automatically attached
when \code{library(r4sub)} is called.
}
\examples{
core_packages()

}
