% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{r4sub_conflicts}
\alias{r4sub_conflicts}
\title{Show R4SUB Function Conflicts}
\usage{
r4sub_conflicts()
}
\value{
A tibble with columns: \code{func}, \code{r4sub_package}, \code{other_package}.
Returned invisibly. Prints conflicts if any exist.
}
\description{
Reports any exported function name collisions between attached R4SUB
packages and other loaded packages.
}
\examples{
r4sub_conflicts()

}
