% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4sub-package.R
\docType{package}
\name{r4sub-package}
\alias{r4sub}
\alias{r4sub-package}
\title{r4sub: Easily Install and Load the R4SUB Ecosystem}
\description{
The 'r4sub' package is a meta-package that installs and loads core packages of the R4SUB clinical submission readiness ecosystem. Loading 'r4sub' attaches 'r4subcore', 'r4subtrace', 'r4subscore', 'r4subrisk', 'r4subdata', and 'r4subprofile'. The dashboard package 'r4subui' is available but not auto-attached.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R4SUB/r4sub}
  \item Report bugs at \url{https://github.com/R4SUB/r4sub/issues}
}

}
\author{
\strong{Maintainer}: Pawan Rama Mali \email{prm@outlook.in}

}
\keyword{internal}
